+++ packages/koot-webpack/libs/new-plugin-workbox-template.js
@@ -1,3 +1,4 @@
+import 'regenerator-runtime/runtime';
 import { setCacheNameDetails } from 'workbox-core';
 import { precacheAndRoute } from 'workbox-precaching';
 import { registerRoute } from 'workbox-routing';